+++ src/pas/plugins/authomatic/__init__.py
@@ -1,9 +1,10 @@
-from AccessControl.Permissions import add_user_folders
-from pas.plugins.authomatic.plugin import AuthomaticPlugin
-from pas.plugins.authomatic.plugin import manage_addAuthomaticPlugin
-from pas.plugins.authomatic.plugin import manage_addAuthomaticPluginForm
-from pas.plugins.authomatic.plugin import tpl_dir
-from Products.PluggableAuthService import registerMultiPlugin
+from pas.plugins.authomatic.patches import apply_patches
+
+
+PACKAGE_NAME = "pas.plugins.authomatic"
+
+
+apply_patches()
 
 
 def initialize(context):
@@ -15,6 +16,13 @@ def initialize(context):
     Here, we call the Archetypes machinery to register our content types
     with Zope and the CMF.
     """
+    from AccessControl.Permissions import add_user_folders
+    from pas.plugins.authomatic.plugin import AuthomaticPlugin
+    from pas.plugins.authomatic.plugin import manage_addAuthomaticPlugin
+    from pas.plugins.authomatic.plugin import manage_addAuthomaticPluginForm
+    from pas.plugins.authomatic.plugin import tpl_dir
+    from Products.PluggableAuthService import registerMultiPlugin
+
     registerMultiPlugin(AuthomaticPlugin.meta_type)
     context.registerClass(
         AuthomaticPlugin,