+++ src/pas/plugins/authomatic/logging.py
@@ -1,0 +1,6 @@
+from pas.plugins.authomatic import PACKAGE_NAME
+
+import logging
+
+
+logger = logging.getLogger(PACKAGE_NAME)